Adding first resource reconciler loop logic
