+++ operator/api/v1alpha1/resourceattestation_types.go
@@ -20,9 +20,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-// EDIT THIS FILE!  THIS IS SCAFFOLDING FOR YOU TO OWN!
-// NOTE: json tags are required.  Any new fields you add must have json tags for the fields to be serialized.
-
 // ResourceAttestationSpec defines the desired state of ResourceAttestation
 type ResourceAttestationSpec struct {
 	ProviderRef AttestationProvider `json:"providerRef"`
