+++ operator/config/samples/_v1alpha1_resource1.yaml
@@ -1,0 +1,15 @@
+apiVersion: argus.io/v1alpha1
+kind: Resource
+metadata:
+  labels:
+    app.kubernetes.io/name: resource
+    app.kubernetes.io/instance: resource-sample
+    app.kubernetes.io/part-of: operator
+    app.kubernetes.io/managed-by: kustomize
+    app.kubernetes.io/created-by: operator
+  name: resource1
+spec:
+  type: my-resource
+  classes:
+  - Myclass
+  parents: []
