+++ operator/config/samples/_v1alpha1_resource2.yaml
@@ -7,6 +7,10 @@ metadata:
     app.kubernetes.io/part-of: operator
     app.kubernetes.io/managed-by: kustomize
     app.kubernetes.io/created-by: operator
-  name: resource-sample
+  name: resource2
 spec:
-  # TODO(user): Add fields here
+  type: my-resource
+  classes:
+  - Myclass
+  parents:
+  - resource2
