+++ operator/internal/controller/implementation/implementation_controller.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controller
+package implementation
 
 import (
 	"context"
@@ -27,26 +27,26 @@ import (
 	argusiov1alpha1 "github.com/ContainerSolutions/argus/operator/api/v1alpha1"
 )
 
-// ResourceAttestationReconciler reconciles a ResourceAttestation object
-type ResourceAttestationReconciler struct {
+// ImplementationReconciler reconciles a Implementation object
+type ImplementationReconciler struct {
 	client.Client
 	Scheme *runtime.Scheme
 }
 
-//+kubebuilder:rbac:groups=argus.io,resources=resourceattestations,verbs=get;list;watch;create;update;patch;delete
-//+kubebuilder:rbac:groups=argus.io,resources=resourceattestations/status,verbs=get;update;patch
-//+kubebuilder:rbac:groups=argus.io,resources=resourceattestations/finalizers,verbs=update
+//+kubebuilder:rbac:groups=argus.io,resources=implementations,verbs=get;list;watch;create;update;patch;delete
+//+kubebuilder:rbac:groups=argus.io,resources=implementations/status,verbs=get;update;patch
+//+kubebuilder:rbac:groups=argus.io,resources=implementations/finalizers,verbs=update
 
 // Reconcile is part of the main kubernetes reconciliation loop which aims to
 // move the current state of the cluster closer to the desired state.
 // TODO(user): Modify the Reconcile function to compare the state specified by
-// the ResourceAttestation object against the actual cluster state, and then
+// the Implementation object against the actual cluster state, and then
 // perform operations to make the cluster state reflect the state specified by
 // the user.
 //
 // For more details, check Reconcile and its Result here:
 // - https://pkg.go.dev/sigs.k8s.io/controller-runtime@v0.14.4/pkg/reconcile
-func (r *ResourceAttestationReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
+func (r *ImplementationReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
 	_ = log.FromContext(ctx)
 
 	// TODO(user): your logic here
@@ -55,8 +55,8 @@ func (r *ResourceAttestationReconciler) Reconcile(ctx context.Context, req ctrl.
 }
 
 // SetupWithManager sets up the controller with the Manager.
-func (r *ResourceAttestationReconciler) SetupWithManager(mgr ctrl.Manager) error {
+func (r *ImplementationReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
-		For(&argusiov1alpha1.ResourceAttestation{}).
+		For(&argusiov1alpha1.Implementation{}).
 		Complete(r)
 }
