+++ operator/internal/controller/requirement/requirement_controller.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controller
+package requirement
 
 import (
 	"context"
@@ -27,26 +27,26 @@ import (
 	argusiov1alpha1 "github.com/ContainerSolutions/argus/operator/api/v1alpha1"
 )
 
-// ImplementationReconciler reconciles a Implementation object
-type ImplementationReconciler struct {
+// RequirementReconciler reconciles a Requirement object
+type RequirementReconciler struct {
 	client.Client
 	Scheme *runtime.Scheme
 }
 
-//+kubebuilder:rbac:groups=argus.io,resources=implementations,verbs=get;list;watch;create;update;patch;delete
-//+kubebuilder:rbac:groups=argus.io,resources=implementations/status,verbs=get;update;patch
-//+kubebuilder:rbac:groups=argus.io,resources=implementations/finalizers,verbs=update
+//+kubebuilder:rbac:groups=argus.io,resources=requirements,verbs=get;list;watch;create;update;patch;delete
+//+kubebuilder:rbac:groups=argus.io,resources=requirements/status,verbs=get;update;patch
+//+kubebuilder:rbac:groups=argus.io,resources=requirements/finalizers,verbs=update
 
 // Reconcile is part of the main kubernetes reconciliation loop which aims to
 // move the current state of the cluster closer to the desired state.
 // TODO(user): Modify the Reconcile function to compare the state specified by
-// the Implementation object against the actual cluster state, and then
+// the Requirement object against the actual cluster state, and then
 // perform operations to make the cluster state reflect the state specified by
 // the user.
 //
 // For more details, check Reconcile and its Result here:
 // - https://pkg.go.dev/sigs.k8s.io/controller-runtime@v0.14.4/pkg/reconcile
-func (r *ImplementationReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
+func (r *RequirementReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
 	_ = log.FromContext(ctx)
 
 	// TODO(user): your logic here
@@ -55,8 +55,8 @@ func (r *ImplementationReconciler) Reconcile(ctx context.Context, req ctrl.Reque
 }
 
 // SetupWithManager sets up the controller with the Manager.
-func (r *ImplementationReconciler) SetupWithManager(mgr ctrl.Manager) error {
+func (r *RequirementReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
-		For(&argusiov1alpha1.Implementation{}).
+		For(&argusiov1alpha1.Requirement{}).
 		Complete(r)
 }
